\name{pbjGUI-package}
\alias{pbjGUI-package}
\alias{pbjGUI}
\docType{package}
\title{
\packageTitle{pbjGUI}
}
\description{
\packageDescription{pbjGUI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pbjGUI}
\packageIndices{pbjGUI}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{pbjGUI}

Maintainer: \packageMaintainer{pbjGUI}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
